+++ elfo-logger/src/actor.rs
@@ -1,8 +1,18 @@
 use std::sync::Arc;
 
-use tokio::{fs::OpenOptions, io::AsyncWriteExt};
+use tokio::{
+    fs::{File, OpenOptions},
+    io::AsyncWriteExt,
+};
 
-use elfo_core::{ActorGroup, Context, Schema};
+use elfo_core as elfo;
+use elfo_macros::{message, msg_raw as msg};
+
+use elfo::{
+    messages::ConfigUpdated,
+    signal::{Signal, SignalKind},
+    ActorGroup, Context, Schema,
+};
 
 use crate::{
     config::{Config, Sink},
@@ -14,6 +24,12 @@ pub(crate) struct Logger {
     ctx: Context<Config>,
     shared: Arc<Shared>,
 }
+
+/// Reload a log file, usually after rotation.
+#[non_exhaustive]
+#[message(elfo = elfo_core)]
+#[derive(Default)]
+pub struct ReopenLogFile {}
 
 impl Logger {
     // TODO: rename it?
@@ -28,29 +44,13 @@ impl Logger {
         Self { ctx, shared }
     }
 
-    async fn main(mut self) {
+    async fn main(self) {
         let mut buffer = String::with_capacity(1024);
+        let mut file = open_file(self.ctx.config()).await;
+        let mut use_colors = can_use_colors(self.ctx.config());
 
-        let use_colors = self.ctx.config().sink == Sink::Stdout && atty::is(atty::Stream::Stdout);
-
-        let config = self.ctx.config();
-        let mut file = if config.sink == Sink::File {
-            // TODO: rely on deserialize instead.
-            let path = config
-                .path
-                .as_ref()
-                .expect("the config path must be provided");
-            Some(
-                OpenOptions::new()
-                    .create(true)
-                    .append(true)
-                    .open(path)
-                    .await
-                    .expect("cannot open the config file"),
-            )
-        } else {
-            None
-        };
+        let signal = Signal::new(SignalKind::Hangup, ReopenLogFile::default);
+        let mut ctx = self.ctx.clone().with(&signal);
 
         // Note that we don't use `elfo::stream::Stream` here intentionally
         // to avoid cyclic dependences (`Context::recv()` logs all messages).
@@ -73,10 +73,15 @@ impl Logger {
                         print!("{}", buffer);
                     }
                 },
-                _envelope = self.ctx.recv() => {
-                    // TODO: logger API.
-                    // TODO: handle SIGHUP.
-                }
+                envelope = ctx.recv() => {
+                    let envelope = ward!(envelope, break);
+                    msg!(match envelope {
+                        ReopenLogFile | ConfigUpdated => {
+                            file = open_file(self.ctx.config()).await;
+                            use_colors = can_use_colors(self.ctx.config());
+                        },
+                    });
+                },
             }
         }
     }
@@ -115,4 +120,29 @@ impl Logger {
 
         out.push('\n');
     }
+}
+
+async fn open_file(config: &Config) -> Option<File> {
+    if config.sink == Sink::Stdout {
+        return None;
+    }
+
+    // TODO: rely on deserialize instead.
+    let path = config
+        .path
+        .as_ref()
+        .expect("the config path must be provided");
+
+    let file = OpenOptions::new()
+        .create(true)
+        .append(true)
+        .open(path)
+        .await
+        .expect("cannot open the config file");
+
+    Some(file)
+}
+
+fn can_use_colors(config: &Config) -> bool {
+    config.sink == Sink::Stdout && atty::is(atty::Stream::Stdout)
 }